+++ tests/detection/test_map.py
@@ -217,6 +217,25 @@ _inputs2 = Input(
     ],
 )
 
+# Test empty preds case, to ensure bool inputs are properly casted to uint8
+# From https://github.com/PyTorchLightning/metrics/issues/981
+_inputs3 = Input(
+    preds=[
+        [
+            dict(boxes=torch.tensor([]), scores=torch.tensor([]), labels=torch.tensor([])),
+        ],
+    ],
+    target=[
+        [
+            dict(
+                boxes=torch.tensor([[1.0, 2.0, 3.0, 4.0]]),
+                scores=torch.tensor([0.8]),
+                labels=torch.tensor([1]),
+            ),
+        ],
+    ],
+)
+
 
 def _compare_fn(preds, target) -> dict:
     """Comparison function for map implementation.
@@ -418,7 +437,7 @@ def _move_to_gpu(input):
 
 @pytest.mark.skipif(_pytest_condition, reason="test requires that torchvision=>0.8.0 is installed")
 @pytest.mark.skipif(_gpu_test_condition, reason="test requires CUDA availability")
-@pytest.mark.parametrize("inputs", [_inputs, _inputs2])
+@pytest.mark.parametrize("inputs", [_inputs, _inputs2, _inputs3])
 def test_map_gpu(inputs):
     """Test predictions on single gpu."""
     metric = MeanAveragePrecision()
@@ -514,7 +533,7 @@ def test_error_on_wrong_input():
         metric.update([], torch.Tensor())  # type: ignore
 
     with pytest.raises(ValueError, match="Expected argument `preds` and `target` to have the same length"):
-        metric.update([dict()], [dict(), dict()])
+        metric.update([{}], [{}, {}])
 
     with pytest.raises(ValueError, match="Expected all dicts in `preds` to contain the `boxes` key"):
         metric.update(